refactor: move ed21559 key pair to new cryptography directory
