+++ src/extension/cryptography/Ed21559KeyPair/index.ts
