+++ src/extension/cryptography/Ed21559KeyPair/types/INewOptions.ts
