+++ src/extension/features/webauthn/thunks/sendWebAuthnAuthenticateResponseThunk.ts
@@ -22,8 +22,8 @@ import { saveAccountsThunk } from '@extension/features/accounts';
 // messages
 import WebAuthnAuthenticateResponseMessage from '@common/messages/WebAuthnAuthenticateResponseMessage';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // types
 import type {
