+++ src/extension/models/PublicKeyCredentialFactory/PublicKeyCredentialFactory.ts
@@ -15,8 +15,8 @@ import { sign } from 'tweetnacl';
 // constants
 import { COSE_ED25519_ALGORITHM } from '@common/constants';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // types
 import type {
