+++ src/extension/models/PublicKeyCredentialFactory/types/IInitOptions.ts
@@ -1,5 +1,5 @@
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // types
 import type { ISerializedPublicKeyCredentialRequestOptions } from '@common/types';
