+++ src/extension/routers/AddAccountMainRouter/AddAccountMainRouter.tsx
@@ -36,8 +36,8 @@ import { create as createNotification } from '@extension/features/notifications'
 // modals
 import AuthenticationModal from '@extension/modals/AuthenticationModal';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // pages
 import AddAccountTypePage from '@extension/pages/AddAccountTypePage';
