+++ src/extension/utils/createAccountImportURI/createAccountImportURI.test.ts
@@ -15,8 +15,8 @@ import {
   ARC0300QueryEnum,
 } from '@extension/enums';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // types
 import type {
