+++ src/extension/utils/fetchDecryptedKeyPairFromStorageWithPasskey/fetchDecryptedKeyPairFromStorageWithPasskey.ts
@@ -4,8 +4,8 @@ import { EncryptionMethodEnum } from '@extension/enums';
 // errors
 import { MalformedDataError } from '@common/errors';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // managers
 import PasskeyManager from '@extension/managers/PasskeyManager';
