+++ src/extension/utils/fetchDecryptedKeyPairFromStorageWithPassword/fetchDecryptedKeyPairFromStorageWithPassword.ts
@@ -6,8 +6,8 @@ import { EncryptionMethodEnum } from '@extension/enums';
 // errors
 import { InvalidPasswordError } from '@common/errors';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // managers
 import PasswordManager from '@extension/managers/PasswordManager';
