+++ src/extension/utils/fetchDecryptedKeyPairFromStorageWithUnencrypted/fetchDecryptedKeyPairFromStorageWithUnencrypted.ts
@@ -1,8 +1,8 @@
 // errors
 import { MalformedDataError } from '@common/errors';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // repositories
 import PrivateKeyRepository from '@extension/repositories/PrivateKeyRepository';
