+++ src/extension/utils/flattenAccountImportSchemaToNewAccounts/flattenAccountImportSchemaToNewAccounts.ts
@@ -3,8 +3,8 @@ import { decodeURLSafe as decodeBase64URLSafe } from '@stablelib/base64';
 // enums
 import { ARC0300QueryEnum } from '@extension/enums';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // types
 import type { IOptions } from './types';
