+++ src/extension/utils/parseURIToARC0300Schema/parseARC0300AccountImportSchema.test.ts
@@ -16,8 +16,8 @@ import {
   ARC0300QueryEnum,
 } from '@extension/enums';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // types
 import type { IARC0300AccountImportSchema } from '@extension/types';
