+++ src/extension/utils/signBytes/signBytes.ts
@@ -6,8 +6,8 @@ import { EncryptionMethodEnum } from '@extension/enums';
 // errors
 import { MalformedDataError } from '@common/errors';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // types
 import type { TOptions } from './types';
