+++ src/extension/utils/signTransaction/signTransaction.ts
@@ -6,8 +6,8 @@ import { EncryptionMethodEnum } from '@extension/enums';
 // errors
 import { MalformedDataError } from '@common/errors';
 
-// models
-import Ed21559KeyPair from '@extension/models/Ed21559KeyPair';
+// cryptography
+import Ed21559KeyPair from '@extension/cryptography/Ed21559KeyPair';
 
 // repositories
 import AccountRepository from '@extension/repositories/AccountRepository';
